refactor : modify the Annotation type NamedEntity, it becomes TextSpan
The category (ie conceot_id) of the TextSpan is dropped , in favour of a
referred Entity in the DatasetItem that could hold a category or have itself
a Classification Annotation associated. This allows a clearer implementation
of multi-modal-annotations (eg VQA with text-image grounding annotations).
A future use-case text centred could require the re-introduction of the
category in the TextSpan , but this is delayed.