fix(back): Return data explicitly in export_dataset_item
Return the updated export_data explicitly in DatasetExporter.export_dataset_item(), instead of using call by sharing to modify it in place, to prevent inconsistencies.

In the case of COCOExporter, we have to edit the existing export_data to merge a BBox with a CompressedRLE if they share the same parent entity, so we cannot just update export_data with append and extend.

For future exporters made by Pixano developers or by users with specific needs, this should also be simpler to understand and to work with.

(Also improve the COCODatasetExporter test coverage)