feat: Add COCO exporter (#462)
* chore(back): Split JSON and JSONL exporters

* feat(back): Add COCO exporter

* feat(back): Add CompressedRLE area property

* feat(back): Update COCOExporter to match COCO JSON format

* chore(back): Fix type hints

* chore(back): Reformat code

* fix(back): Return data explicitly in export_dataset_item

Return the updated export_data explicitly in DatasetExporter.export_dataset_item(), instead of using call by sharing to modify it in place, to prevent inconsistencies.

In the case of COCOExporter, we have to edit the existing export_data to merge a BBox with a CompressedRLE if they share the same parent entity, so we cannot just update export_data with append and extend.

For future exporters made by Pixano developers or by users with specific needs, this should also be simpler to understand and to work with.

(Also improve the COCODatasetExporter test coverage)

* chore(back): Fix type hints

* refactor(back): Simplify export_dataset_item loop

- Checking if value is either None or an empty list can be combined as both None and [] evaluate as False
- Rather than duplicating code for processing the schema_data as a list or as a single value, make lists out of single values.

* refactor(back): Update export_dataset_item

* feat(back): Add categories in COCOExporter

- Add "category_name" if category field in Entity
- Add "category_id" if category_name in category_dict (either a default dictionary like "coco91", "coco80", and "voc", or a custom user-provided dictionary)

* chore(back): Fix type hints

* fix(back): Check the correct ID for existing_coco_ann of entities

Also reformat code

* chore(back): Add missing space from code reformarting