feat(front): Improve object highlighting (#449)
* feat(front): Remove current scrolling from ObjectInspector

Only scrolls to top to show the selected thumbnails, not as useful as scrolling to the object in the list

* feat(front): Remove current scrolling VideoInspector

* feat(front): Remove createObjectCardId function

Not necessary to have a separate function for this single use.
A similar action is performed in ObjectTrack without using a function.

* feat(front): Add object highlighting for both cards and tracks

Add highlightObject function in objectsApi that highlights and scrolls into view both the ObjectCard and the ObjectTrack.
Still works on image datasets that don't use the ObjectTrack.

Add buttons in ObjectTracks to trigger this highlighting.

* chore(front): Remove unused import

* fix(front): Scroll into vew regardless of highlightFrameIndex

The lines for scrolling into view in the VideoInspector and ObjectsInspector were inside the if statement for changing the currentFrameIndex by mistake.

* fix(front): Go to highlightFrameIndex only if object not visible

If the highlighted object is already visible (i.e. the current frame index is between the start and end timesteps of at least one tracklet), there is no need to move the currentFrameIndex as the object is already visible.

* fix(front): Fix highlightFrameIndex conditions

* fix(front): Fix smooth scrolling on highlighting

Replace  `behavior: "smooth"` of scrollIntoView that is not supported in all browsers with Tailwind `scroll-smooth` property

* remove double click, highlight on canvas single click (rectangle)

* fix(front): Use svelte/store get in highlightObject

* remove update of currentFrameIndex for canvas selection + use store.get for entities in getTopEntity

* remove log

* correctly update view + fix highlights

* lint

* fix wrong track highlight + minor refacto

* typo

* dont scroll on unselect + scroll in fusion too

* better support for Image workspace + keypoints + mask

* fix(front): Single scrollbar in ObjectsInspector

* disable both dot click in fusion mode + never scroll when deselect

---------

Co-authored-by: brenault <bertrand.renault@cea.fr>