feat(front): Disable changing frame index on right click
When editing tracklets in long videos, having the cursor move on the right click is difficult to work with, as it requires being very precise and/or zooming very closely on the tracklet to right click at the exact position. It is much easier to get to the exact position where you want to perform an action (like "Split tracklet or "Add a point") by using the video controls (Play / Pause, Previous frame, Next frame).

This commit disables changing the frame index on right click, and actions like "Split tracklet" or "Add a point" are instead performed on the current frame index if possible. If the frame index is not compatible with those actions, they are hidden.