fix(front): Minor UI fixes (#515)
* fix(front): Change text depending on number of objects

* fix(front): Fix higlighting outside of fusion tool

Making objects with highlight state equal to "none" more transparent is really only necessary in fusion tool, for objects that cannot be selected because of overlap.

For regular highlighting, this not really needed, especially now that the highlighted object is much more visible with both the updated scroll into view, and the higher opacity colors.